+++ src/Commands/GreetCommand.php
@@ -1,0 +1,75 @@
+<?php
+
+namespace Rougin\Blueprint\Commands;
+
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+/**
+ * Greet Command
+ *
+ * A sample of a greet command.
+ *
+ * @package Blueprint
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class GreetCommand extends AbstractCommand
+{
+    /**
+     * Checks whether the command is enabled or not in the current environment.
+     *
+     * @return boolean
+     */
+    public function isEnabled()
+    {
+        $defined  = defined('BLUEPRINT_FILENAME');
+        $filename = 'blueprint.yml';
+
+        return file_exists($defined ? BLUEPRINT_FILENAME : $filename);
+    }
+
+    /**
+     * Sets the configurations of the current command.
+     *
+     * @return void
+     */
+    protected function configure()
+    {
+        $this
+            ->setName('greet')
+            ->setDescription('Greet someone')
+            ->addArgument(
+                'name',
+                InputArgument::OPTIONAL,
+                'Who do you want to greet?'
+            )
+            ->addOption(
+                'yell',
+                null,
+                InputOption::VALUE_NONE,
+                'If set, the task will yell in uppercase letters'
+            )
+        ;
+    }
+
+    /**
+     * Executes the current command.
+     *
+     * @param  \Symfony\Component\Console\Input\InputInterface  $input
+     * @param  \Symfony\Component\Console\Input\OutputInterface $output
+     * @return void|string
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $name = $input->getArgument('name') ?: 'Stranger';
+        $text = 'Hello ' . $name . '!';
+
+        if ($input->getOption('yell')) {
+            $text = strtoupper($text);
+        }
+
+        return $output->writeln($text);
+    }
+}